Deps: Update flutter_secure_storage and related packages
This commit updates the `flutter_secure_storage` package from version `4.2.1` to `9.2.4`.

The `pubspec.lock` file has been updated to reflect this change and the new versions of its transitive dependencies:
- `flutter_secure_storage_linux`: `1.2.3`
- `flutter_secure_storage_macos`: `3.1.3`
- `flutter_secure_storage_platform_interface`: `1.1.2`
- `flutter_secure_storage_web`: `1.2.1`
- `flutter_secure_storage_windows`: `3.1.2`

Additionally, the `js` package was updated from `0.7.2` to `0.6.7` as a result of dependency resolution.

The `pubspec.yaml` file was also updated to reflect the new `flutter_secure_storage` version.